# api/openapi.yaml
openapi: 3.0.3
info:
  title: PGDock API
  version: 1.0.0
  description: |
    Control-plane API for PGDock. This file is the single source of truth:
    Go server interfaces (oapi-codegen) and TypeScript client types
    (openapi-typescript) are generated from it with `make generate`.
servers:
  - url: /
paths:
...
  /api/v1/operations:
    get:
      operationId: listOperations
      summary: List operations, newest first
      tags: [operations]
      parameters:
        - $ref: "#/components/parameters/OrgQuery"
        - name: platform
          in: query
          description: Platform-level operations (no project), for the platform admin.
          schema:
            type: boolean
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/OperationStatus"
        - name: kind
          in: query
          schema:
            type: string
        - name: project_id
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Operations matching the filters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationList"
        default:
          $ref: "#/components/responses/Error"
...
  /api/v1/projects:
    get:
      operationId: listProjects
      summary: List projects you can see in an organisation, newest first
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/OrgQuery"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ProjectStatus"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: Live projects.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectList"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createProject
      summary: Create a shared-tier project
      description: |
        Records the project and queues a `create` operation. The response
        carries the password and full connection URLs exactly once; PGDock
        stores only the SCRAM verifier. The URLs work once the operation
        succeeds (stream it to follow progress).
      tags: [projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "202":
          description: Project recorded and create operation queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectCredentials"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/projects/{id}:
    get:
      operationId: getProject
      summary: Project detail and connection info (no password)
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: The project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteProject
      summary: Delete a project
      description: |
        `confirm` must equal the project's name, and the session must have
        re-authenticated recently (`POST /auth/reauth`); otherwise 403 with
        code `reauth_required`. Queues a `delete` operation that takes a
        final backup (kept 30 days; `skip_final_backup` skips it), removes
        the pooler route, and drops the database and role.
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - name: confirm
          in: query
          required: true
          schema:
            type: string
        - name: skip_final_backup
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "202":
          description: Delete operation queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        default:
          $ref: "#/components/responses/Error"
...
  /api/v1/projects/{id}/rotate-password:
    post:
      operationId: rotateProjectPassword
      summary: Rotate the project password
      description: |
        Returns the new password once and queues a `rotate` operation. The
        new password works, and the old one stops working, when it succeeds.
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "202":
          description: Rotate operation queued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectCredentials"
        default:
          $ref: "#/components/responses/Error"
...
components:
  parameters:
...
    OrgQuery:
      name: org
      in: query
      description: The organisation to list; your personal organisation when omitted.
      schema:
        type: string
        format: uuid
...
    ProjectID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: not_implemented
        message:
          type: string
        quota:
          $ref: "#/components/schemas/QuotaItem"
    OperationStatus:
      type: string
      enum: [queued, running, succeeded, failed]
    OperationLogEntry:
      type: object
      required: [ts, step, level, msg]
      properties:
        ts:
          type: string
          format: date-time
        step:
          type: string
        level:
          type: string
          enum: [info, warn, error]
        msg:
          type: string
    Operation:
      type: object
      required: [id, kind, params, status, attempts, run_after, log, created_at]
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
          example: create
        project_id:
          type: string
          format: uuid
          nullable: true
        params:
          type: object
          additionalProperties: true
        status:
          $ref: "#/components/schemas/OperationStatus"
        attempts:
          type: integer
        run_after:
          type: string
          format: date-time
        log:
          type: array
          items:
            $ref: "#/components/schemas/OperationLogEntry"
        error:
          type: string
          nullable: true
        created_by:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
          nullable: true
    OperationList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Operation"
...
    ProjectStatus:
      type: string
      enum: [provisioning, active, promoting, restoring, deleting, error]
    ProjectTier:
      type: string
      enum: [shared, dedicated]
    ConnectionInfo:
      type: object
      required: [host, session_port, pooled_port, database, user, sslmode, pooled_url, session_url]
      properties:
        host:
          type: string
          example: db.example.com
        session_port:
          type: integer
          example: 5432
        pooled_port:
          type: integer
          example: 6543
        database:
          type: string
          example: blog_k2f9
        user:
          type: string
          example: blog_k2f9_owner
        sslmode:
          type: string
          example: require
        pooled_url:
          type: string
          description: Transaction-mode URL (app traffic). Without a password unless returned at creation or rotation.
          example: postgresql://blog_k2f9_owner@db.example.com:6543/blog_k2f9?sslmode=require
        session_url:
          type: string
          description: Session-mode URL (migrations). Without a password unless returned at creation or rotation.
          example: postgresql://blog_k2f9_owner@db.example.com:5432/blog_k2f9?sslmode=require
    Project:
      type: object
      required: [id, org_id, name, slug, db_name, owner_role, tier, status, settings, created_at, connection]
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        my_role:
          $ref: "#/components/schemas/ProjectRole"
        storage_state:
          $ref: "#/components/schemas/StorageState"
        can_switch_credentials:
          type: boolean
          description: A V1 project that still uses its V1 owner role (V2 §10.2).
        legacy_credentials_until:
          type: string
          format: date-time
          nullable: true
          description: The V1 credentials stop working at this time.
        name:
          type: string
        slug:
          type: string
        db_name:
          type: string
        owner_role:
          type: string
        tier:
          $ref: "#/components/schemas/ProjectTier"
        status:
          $ref: "#/components/schemas/ProjectStatus"
        description:
          type: string
          nullable: true
        settings:
          $ref: "#/components/schemas/ProjectSettings"
        created_at:
          type: string
          format: date-time
        connection:
          $ref: "#/components/schemas/ConnectionInfo"
        last_backup_at:
          type: string
          format: date-time
          nullable: true
          description: When the latest backup of this project finished.
        instance:
          $ref: "#/components/schemas/InstanceSummary"
        pitr_window:
          $ref: "#/components/schemas/PitrWindow"
        retired_copy_until:
          type: string
          format: date-time
          nullable: true
          description: After a promotion, when the read-only shared copy is dropped.
    ProjectSettings:
      type: object
      required: [connection_limit, pool_size, statement_timeout, idle_in_transaction_session_timeout, disk_warn_bytes, console_read_only]
      properties:
        connection_limit:
          type: integer
        pool_size:
          type: integer
        statement_timeout:
          type: string
        idle_in_transaction_session_timeout:
          type: string
        disk_warn_bytes:
          type: integer
          format: int64
        console_read_only:
          type: boolean
    ProjectList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Project"
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        org_id:
          type: string
          format: uuid
          description: The organisation; your personal organisation when omitted.
        name:
          type: string
          minLength: 1
          maxLength: 64
          example: My Blog
        description:
          type: string
          maxLength: 1000
        tier:
          $ref: "#/components/schemas/ProjectTier"
        node_id:
          type: string
          format: uuid
          description: Dedicated only; default is the least loaded dedicated node.
        profile:
          type: string
          description: Dedicated only (see /profiles); default small.
        volume_gb:
          type: integer
          minimum: 1
          maximum: 16384
          description: Dedicated only; default 20.
    ProjectCredentials:
      type: object
      description: Shown once. PGDock keeps only the SCRAM verifier.
      required: [project, operation, password, connection]
      properties:
        project:
          $ref: "#/components/schemas/Project"
        operation:
          $ref: "#/components/schemas/Operation"
        password:
          type: string
        connection:
          $ref: "#/components/schemas/ConnectionInfo"
...
    InstanceSummary:
      type: object
      required: [id, kind, status, node_id, node_name]
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
          enum: [shared, dedicated]
        status:
          type: string
        error:
          type: string
          nullable: true
        node_id:
          type: string
          format: uuid
        node_name:
          type: string
        profile:
          type: string
          nullable: true
        cpus:
          type: number
          nullable: true
        memory_mb:
          type: integer
          nullable: true
        volume_gb:
          type: integer
          nullable: true
    PitrWindow:
      type: object
      description: Dedicated only. Any time in [from, to] can be restored.
      required: [from, to]
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
...
    ProjectRole:
      type: string
      enum: [admin, developer, read_only]
...
    QuotaItem:
      type: object
      required: [limit, used]
      properties:
        limit:
          type: string
          description: The limit's key (V2 §10.3), e.g. projects or project_storage_mb.
        used:
          type: number
        max:
          type: integer
          format: int64
          nullable: true
          description: Absent or null means unlimited.
...
    StorageState:
      type: string
      description: V2 §10.4 storage enforcement.
      enum: [none, warn, soft, hard]
...